// react-frontend/src/api/models/select-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SelectDTO
 */
export interface SelectDTO {
    /**
     * 
     * @type {number}
     * @memberof SelectDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof SelectDTO
     */
    'name'?: string;
}